Version validation on the command line
dub will now build the semver application with the first command line usage.

Given the -v or --validate option and one version string, the application will return 1 if the version string is not compliant to the Semantic Versioning specification. If the version string is okay, 0 is returned.

In order to test the output of the command line tool and to keep cli interface in sync with the future library interface, the test folder has been created. It contains a test script for testing the command line tool and the expectations of the version validation.