Declaring public API | function safety
The Semantic versioning sepecification requires the declaration of a public API. I decided to just take the API of semver.d into account, since the command line tool is just using it and I plan to build static libraries from semver.d later on.

To make the API obvious, I provided the public interface in a separate 'public' block. With aliasing, I also gave them better names than the ones used inside.

I also had a look at safety for the function and changed it to be safe. I'd rather choose to make it a pure function, but std.regex.match is not pure.